% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_functions.R
\name{asymcov.gmm}
\alias{asymcov.gmm}
\title{estimate asymptotic covariance of the GMM estimator}
\usage{
asymcov.gmm(eta, sigmasqBk, sigmasqY, cov.mat, stage2.conduct = T)
}
\description{
estimate asymptotic covariance of the GMM estimator
}
