% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM_functions.R
\name{cal.Omega}
\alias{cal.Omega}
\title{covariance matrix of the estimating equations}
\usage{
cal.Omega(eta, sigmasqBk, sigmasqY, cov.mat)
}
\description{
covariance matrix of the estimating equations
}
