% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat.R
\name{covinfo}
\alias{covinfo}
\title{Provide information on between-biomarker correlation matrix.}
\usage{
covinfo(
  sumstats,
  out.path,
  ldsc.path,
  python.path = NULL,
  ldscore.path = file.path(ldsc.path, "eur_w_ld_chr/"),
  maf.thr,
  mergeallele = TRUE,
  K
)
}
\arguments{
\item{sumstats}{a list of matrices storing the GWAS summary-level association statistics of
the biomarkers.}

\item{out.path}{path to the folder where the output from LD score regression (LDSC) will be saved.}

\item{ldsc.path}{path to the folder where the required files for LDSC are stored.}

\item{python.path}{path to the python software.}

\item{ldscore.path}{path to the ldscore files. Default: file.path(ldsc.path,"eur_w_ld_chr/").}

\item{mergeallele}{a logical constant for whether or not the alleles should be merged.
Default: TRUE.}
}
\value{
a list containing (1) coherit.mat: the estimated genetic correlation matrix;
(2) ldscint.mat: the calculated LD score regression intercept matrix.
}
\description{
Provide information on between-biomarker correlation matrix.
}
