% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat.R
\name{covmat}
\alias{covmat}
\title{second derivatives}
\usage{
covmat(
  sumstats.all,
  traitvec,
  out.path,
  ldsc.path,
  python.path = NULL,
  ldscore.path = file.path(ldsc.path, "eur_w_ld_chr/"),
  mergeallele = TRUE
)
}
\description{
second derivatives
}
