% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrle.R
\name{mrle}
\alias{mrle}
\title{Mendelian Randomization analysis using multiple Biomarkers of an underlying common exposure}
\usage{
mrle(sumtable, thetak_sign, Cov.mat, alpha0 = 0.05)
}
\arguments{
\item{sumtable}{the M by (K+1)*3+6 input matrix containing}

\item{thetak_sign}{a vector of length K indicating the directions of the effect of the latent exposure on
the traits/biomarkers}

\item{Cov.mat}{the (K+1) by (K+1) estimated covariance matrix of the GWAS summary statistics.
The first variable corresponds to the outcome; the 2nd to the (K+1)-th variables correspond to
the k-th observable traits.}

\item{alpha0}{significance level of the hypothesis test.}
}
\value{
A matrix of the infile
}
\description{
This function conducts Mendelian Randomization (MR) analysis for a latent exposure (MRLE).
The method uses summary-level association statistics from genome-wide association study
of the outcome and K>=3 observable traits/biomarkers that are co-regulated by the latent exposure
on a set of SNPs (instrumental variables) that are associated with
at least two of the traits/biomarkers at specified thresholds of significance.
}
