% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing}
\usage{
preprocess(sumstats, maf.thr, K)
}
\arguments{
\item{sumstats}{a list of matrices storing the GWAS summary-level association statistics of the biomarkers.
Each matrix contains the following columns: rsid, Chr, Pos, A2, A1, beta, se, P, N.}

\item{maf.thr}{MAF threshold used for filtering out SNPs that have MAF<maf.thr.}

\item{K}{number of biomarkers used in the analysis (K>=3).}
}
\value{
a list containing (1) sumstats.all: a matrix that contains the merged summary data;
(2) traitvec: a vector storing the names of the biomarkers
}
\description{
Data preprocessing
}
