% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preprocessing.R
\name{sumdata_merge}
\alias{sumdata_merge}
\title{Merge GWAS summary-level association statistics of two traits}
\usage{
sumdata_merge(
  sumdata1,
  sumdata2,
  A1.col.1,
  A2.col.1,
  A1.col.2,
  A2.col.2,
  beta.col.2
)
}
\arguments{
\item{sumdata1}{a matrix storing GWAS summary-level association statistics of the first trait (name: traitname1)
that contain columns 'loc', 'Chr', 'Pos', 'rsid', A1.col.1, A2.col.1, and paste0(c('beta', 'se', 'P'), '.', traitname1).
Here, A1.col.1, A2.col.1, beta, se and P represent allele1, allele2, the estimated effect size
(corresponding to allele1), standard error and p-value obtained from GWAS.}

\item{sumdata2}{a matrix storing GWAS summary-level association statistics of the first trait (name: traitname2)
with columns A1.col.2, A2.col.2, 'rsid', and paste0(c('beta', 'se', 'P'), '.', traitname2).
Here, A1.col.1, A2.col.1, beta, se and P represent allele1, allele2, the estimated effect size
(corresponding to allele1), standard error and p-value obtained from GWAS.}

\item{A1.col.1}{column name in sumdata1 that corresponds to allele1 of the summary data for the first trait.}

\item{A2.col.1}{column name in sumdata1 that corresponds to allele2 of the summary data for the first trait.}

\item{A1.col.2}{column name in sumdata1 that corresponds to allele1 of the summary data for the second trait.}

\item{A2.col.2}{column name in sumdata1 that corresponds to allele2 of the summary data for the second trait.}

\item{beta.col.2}{column name in sumdata2 that corresponds to beta of the summary data for the second trait.}
}
\value{
Merged summary data matrix
}
\description{
This function merge GWAS summary-level association statistics of two traits according to rsid.
It then matches alleles of the second trait to those of the first trait with the following steps:
\enumerate{
\item Align flipped alleles
\item Align alleles with modifiable strand-ambiguous
}
}
